refactor(mm-next/photography): move HeroImage into photo-slider component
